feat: move KnowledgeToFeedInput to @lib
