+++ frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/KnowledgeToFeed.tsx
@@ -3,12 +3,12 @@ import { useTranslation } from "react-i18next";
 import { MdClose } from "react-icons/md";
 
 import { AddBrainModal } from "@/lib/components/AddBrainModal";
+import { KnowledgeToFeedInput } from "@/lib/components/KnowledgeToFeedInput";
 import Button from "@/lib/components/ui/Button";
 import { Select } from "@/lib/components/ui/Select";
 import { requiredRolesForUpload } from "@/lib/config/upload";
 import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
 
-import { KnowledgeToFeedInput } from "./components/KnowledgeToFeedInput";
 import { formatMinimalBrainsToSelectComponentInput } from "./utils/formatMinimalBrainsToSelectComponentInput";
 
 type KnowledgeToFeedProps = {
