--- frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/components/KnowledgeToFeedInput/index.ts
@@ -1,1 +1,0 @@
-export * from "./KnowledgeToFeedInput";
